#61 added draft cargo

+++ mechanics-module/src/main/java/marowak/dev/dto/equipment/Cargo.java
@@ -1,4 +1,4 @@
 package marowak.dev.dto.equipment;
 
-public record Cargo(int id, int slotId) {
+public record Cargo(long id, int slotId) {
 }
